Sonarcloud updates (#74)
* Sonarcloud updates

* further sonarcloud updates

* further sonarcloud updates

+++ coherence-visualvm-plugin/src/main/java/com/oracle/coherence/plugin/visualvm/helper/GraphHelper.java
@@ -307,7 +307,7 @@ public class GraphHelper
      */
     public static void addValuesToTotalProxyConnectionsGraph(SimpleXYChartSupport graph, int cTotalConnections)
         {
-        graph.addValues(System.currentTimeMillis(), new long[] {(long) cTotalConnections});
+        graph.addValues(System.currentTimeMillis(), new long[] {cTotalConnections});
         }
 
     /**
